Camera Bounds, Mouse Scroll and Hex Highlighting
Placed adjustable bounds on the camera in the CombatMap within the
Position class.
Added functionality to scroll the camera by hovering the mouse over the
edge of the application window.
Added a highlight on the CombatMap to the hex which the cursor is
hovering over.
Some initial structure for Agent and ActiveAgent classes.

Removed the unused Graphics class.