Initial CombatMap UI
Implemented basic (visual) UI to the CombatMap in the form of a menu
across the bottom of the screen.
Adjusted camera bounds to allow for the entire map to be seen with the
addition of the UI.
Added SetDimensions function to the Bitmap class to allow for resizing
of a bitmap (and its texture) without reinitialization or the use of a
second instance.
Repositioned the return to main menu functionality on the CombatMap to
the bottom right button on the UI, providing a visual marker for the
functionality.