Agent Selection
Left clicking on a hex with an Agent selects that Agent. In the event of
multiple Agents being in the same hex, the first one checked is
selected. If an empty hex is clicked on, any selected Agent is
deselected.
Right clicking on a hex now moves the currently selected Agent to that
hex instead of moving the first Agent.
Removed the "Goodbye" string.
Added a String that is displayed on the top left of the Combat Map menu
bar denoting which Agent is currently selected, or nothing if no Agent
is selected.