Turn System Proof of Concept
Implemented functionality for the End Turn button on the CombatMap menu
bar.
The End Turn button ends the turn of all currently active Agents and
begins next turn. When all Agents have ended their turn in a round, a
new round begins.
Added a bounds check when updating the strings that display the cursor
position to eliminate a fatal exception with unsafe values.